--- src/Config.js
@@ -1,2 +1,0 @@
-const url = "http://10.58.4.214:8000";
-export default url;
