+++ src/configBasket.js
@@ -1,0 +1,3 @@
+const urlBasket = "http://10.58.0.34:8000";
+
+export default urlBasket
