+++ src/configMenu.js
@@ -1,0 +1,2 @@
+const urlMenu = "http://10.58.0.78:8000";
+export default urlMenu;
